Fix snippet remapping IPs for NAT setup
For the setup with a NAT, the Bind address for the border router needs to
be the internal IP, but the Public IP in the topology file has to be
the external IP and shouldn't be changed.
The snippet was previously replacing all occurrences of the external IP
with the internal one. The new snippet only replaces the address in the
Bind-field.