Update/simplify SIG tutorial
Refer to new package installation instead of build instructions.

Simplify example setup; no dummy interfaces, use `ip route` instead of
`ip rule`.
This setup is clearly less "powerfull", i.e. it might not be clear how
to go from here to a setup with routing.
Clearly, having the simple version first is friendlier -- not sure
wheter the "extended" version with routing with all the dummy interfaces
needs be part of this guide.